+++ .gitignore
@@ -177,3 +177,4 @@ todolist.txt
 .pypirc
 >>>>>>> upstream/main
 slitmaskgui/tests/notes.txt
+notes.txt
