+++ slitmaskgui/tests/test_mask_gen.py
@@ -1,61 +1,54 @@
-from slitmaskgui.backend.mask_gen import SlitMask
+# from slitmaskgui.backend.mask_gen import SlitMask
 
-import numpy as np
-from astropy.coordinates import SkyCoord
-import astropy.units as u
+
+import pytest
+from slitmaskgui.backend.mask_gen import SlitMask, CSU_HEIGHT, CSU_WIDTH, TOTAL_BAR_PAIRS
 """
 will be in a list of obj {name,ra,dec,equinox,vmag,priority,bar_id,x_mm,y_mm}
 """
 
-def generate_random_center():
 
-    ra_random = np.random.uniform(0, 360)  # RA in degrees
-    dec_random = np.random.uniform(-90, 90)  # Dec in degrees
-
-    random_coord = SkyCoord(ra=ra_random, dec=dec_random, unit='deg', frame='icrs')
-
-    ra_ = random_coord.ra.to_string(unit='hour', sep=' ', precision=2)
-    dec_ = random_coord.dec.to_string(unit='deg', sep=' ', precision=2)
-    center = f"{ra_} {dec_}"
-    print(center)
-    center_coord = SkyCoord(ra=ra_, dec=dec_, unit=(u.hourangle,u.deg), frame='icrs')
-    return center_coord
-
-center = generate_random_center()
+# print(stars)
 
 
-def make_bunch_o_stars(center, radius, num_stars):
-    star_list = []
-    center_ra = center.ra.deg
-    center_dec = center.dec.deg
+def test_check_if_within_bounds():
+    sm = SlitMask([])
 
+    assert sm.check_if_within(0, 0) is True
+    assert sm.check_if_within(CSU_WIDTH/2 - 0.01, CSU_HEIGHT/2 - 0.01) is True
+    assert sm.check_if_within(CSU_WIDTH/2 + 1, 0) is False
+    assert sm.check_if_within(0, CSU_HEIGHT/2 + 1) is False
 
-    for _ in range(num_stars):
-        # Generate a random offset within a circle (uniform in area)
+def test_bar_id_assignment_center():
+    stars = [{"x_mm": 0, "y_mm": 0, "priority": 1}]
+    mask = SlitMask(stars)
+    assert "bar_id" in mask.stars[0]
+    assert mask.stars[0]["bar_id"] == TOTAL_BAR_PAIRS // 2
 
-        rand_radius = radius * np.sqrt(np.random.uniform(0, 1))  # sqrt for uniform density
-        rand_angle = np.random.uniform(0, 2 * np.pi)
+def test_out_of_bounds_star_is_removed():
+    stars = [
+        {"x_mm": 0, "y_mm": 0, "priority": 1},
+        {"x_mm": CSU_WIDTH + 1, "y_mm": 0, "priority": 2},  # Should be removed
+    ]
+    mask = SlitMask(stars)
+    assert len(mask.stars) == 1
+    assert mask.stars[0]["x_mm"] == 0 and mask.stars[0]["y_mm"] == 0
 
-        # Offset in RA/Dec (approximation valid for small radius)
-        delta_ra = (rand_radius * np.cos(rand_angle)) / np.cos(np.deg2rad(center_dec))
-        delta_dec = rand_radius * np.sin(rand_angle)
+def test_priority_optimization_per_bar_id():
+    stars = [
+        {"x_mm":0,"y_mm":1, "priority": 1},
+        {"x_mm":0,"y_mm":1, "priority": 3},  # same bar_id, higher priority
+        {"x_mm":0,"y_mm":-1, "priority": 2}, #  different bar_id, does not pass if y_mm is positive 40
+    ]
+    mask = SlitMask(stars)
+    result = mask.return_mask()
 
-        # Calculate new coordinates
-        new_ra = center_ra + delta_ra
-        new_dec = center_dec + delta_dec
+    # Check: 1 star per bar_id
+    bar_ids = [s["bar_id"] for s in result]
+    print(bar_ids)
+    assert len(bar_ids) == len(result)
 
-        # Wrap RA to [0, 360) and clamp Dec to [-90, 90]
-        new_ra = new_ra % 360
-        new_dec = max(min(new_dec, 90), -90)
-
-        star_coord = SkyCoord(ra=new_ra * u.deg, dec=new_dec * u.deg, frame='icrs')
-
-        ra_str = star_coord.ra.to_string(unit='hour', sep=' ', precision=2, pad=True)
-        dec_str = star_coord.dec.to_string(unit='deg', sep=' ', precision=2, alwayssign=True)
-
-        priority = int(np.random.uniform(1,2000))
-        star_list.append({"ra":ra_str, "dec": dec_str,"priority":priority})
-    return star_list
-    
-stars = make_bunch_o_stars(center,radius=10/60,num_stars=100)
-print(stars)
+    # Check: highest priority per group is kept
+    for star in result:
+        if star["bar_id"] == mask.stars[0]["bar_id"]:
+            assert star["priority"] == 3
